+++ trading_framework/execution_client.py
@@ -2,7 +2,11 @@ from typing import Protocol
 
 
 class ExecutionException(Exception):
-    pass
+    """Exception raised for errors in the execution of orders."""
+
+    def __init__(self, message):
+        self.message = message
+        super().__init__(self.message)
 
 
 class ExecutionClient(Protocol):
@@ -14,7 +18,10 @@ class ExecutionClient(Protocol):
         :param amount: the amount to buy
         :return: None
         """
-        ...
+        try:
+            print("Executing buy order for {0} shares of {1}".format(amount,product_id))
+        except ExecutionException as e:
+            print("An error occurred:", e)
 
     def sell(self, product_id: str, amount: int):
         """
@@ -23,4 +30,7 @@ class ExecutionClient(Protocol):
         :param amount: the amount to sell
         :return: None
         """
-        ...
+        try:
+            print("Executing sell order for {0} shares of {1}".format(amount,product_id))
+        except ExecutionException as e:
+            print("An error occurred:", e)
