+++ trading_framework/price_listener.py
@@ -10,4 +10,6 @@ class PriceListener(Protocol):
         :param price: the current market price of hte product
         :return: None
         """
-        ...
+
+        print(f"Received price tick for product {product_id}: ${price}")
+
